@@ -3,7 +3,7 @@ source 'https://rubygems.org'
 # Specify your gem's dependencies in liquid-rails.gemspec
 gemspec
 
-gem 'rails',       '~> 4.2.10'
+gem 'rails',       '~> 5.0.6'
 gem 'rake',        '~> 12.3.0'
 gem 'rspec-rails', '~> 3.5.2'
 gem 'guard-rspec', '~> 4.7.3'
@@ -11,3 +11,4 @@ gem 'capybara',    '~> 2.16.1'
 gem 'pry-rails',   '~> 0.3.6'
 gem 'coveralls',   '~> 0.8.21'
 gem 'simplecov',   '~> 0.14.1'
+gem 'sqlite3',     '~> 1.3.13'
