@@ -20,7 +20,7 @@ Gem::Specification.new do |spec|
   spec.test_files    = spec.files.grep(%r{^(test|spec|features)/})
   spec.require_paths = ['lib']
 
-  spec.add_dependency 'rails',    ">= 3.2", "< 5"
-  spec.add_dependency 'liquid',   '~> 3.0.6'
-  spec.add_dependency 'kaminari', '~> 0.17.0'
+  spec.add_dependency 'rails',    ">= 5.0.6"
+  spec.add_dependency 'liquid',   '~> 4.0.0'
+  spec.add_dependency 'kaminari', '~> 1.1.1'
 end
