@@ -1,5 +1,5 @@
 class HomeController < ApplicationController
-  before_filter :set_book
+  before_action :set_book
 
   def index
     render layout: false
@@ -20,9 +20,13 @@ class HomeController < ApplicationController
   def erb_with_html_liquid_partial
   end
 
+  def index_with_rss
+    render layout: false
+  end
+
   private
 
-    def set_book
-      @book = { 'name' => 'Liquid on Rails' }
-    end
+  def set_book
+    @book = { 'name' => 'Liquid on Rails' }
+  end
 end
