@@ -6,9 +6,9 @@ module Liquid
 
         def setup_view_and_controller
           @view                 = ActionView::Base.new
-          @controller           = ApplicationController.new
-          @request              = ActionController::TestRequest.new('PATH_INFO' => '/')
-          @response             = ActionController::TestResponse.new
+          @controller           = ActionController::Base.new
+          @request              = ActionDispatch::TestRequest.new({'PATH_INFO' => '/'})
+          @response             = ActionDispatch::TestResponse.new
           @response.request     = @request
           @controller.request   = @request
           @controller.response  = @response
