@@ -13,8 +13,6 @@ module Liquid
       end
 
       def render(template, local_assigns={})
-        @view.controller.headers['Content-Type'] ||= 'text/html; charset=utf-8'
-
         assigns = if @controller.respond_to?(:liquid_assigns, true)
           @controller.send(:liquid_assigns)
         else
@@ -23,9 +21,8 @@ module Liquid
         assigns['content_for_layout'] = @view.content_for(:layout) if @view.content_for?(:layout)
         assigns.merge!(local_assigns.stringify_keys)
 
-        liquid = Liquid::Template.parse(template)
-        render_method = (::Rails.env.development? || ::Rails.env.test?) ? :render! : :render
-        liquid.send(render_method, assigns, filters: filters, registers: { view: @view, controller: @controller, helper: @helper }).html_safe
+        liquid      = Liquid::Template.parse(template)
+        liquid.send(render_method, assigns, filters: filters, registers: registers).html_safe
       end
 
       def filters
@@ -36,8 +33,21 @@ module Liquid
         end
       end
 
+      def registers
+        {
+          view: @view,
+          controller: @controller,
+          helper: @helper,
+          file_system: Liquid::Rails::FileSystem.new(@view)
+        }
+      end
+
       def compilable?
         false
+      end
+
+      def render_method
+        (::Rails.env.development? || ::Rails.env.test?) ? :render! : :render
       end
     end
   end
