@@ -1,5 +1,5 @@
 module Liquid
   module Rails
-    VERSION = '0.1.4'
+    VERSION = '0.2.0.beta1'
   end
 end
