@@ -1,7 +1,7 @@
-require File.expand_path('../boot', __FILE__)
+require_relative 'boot'
 
 # Pick the frameworks you want:
-require "action_controller/railtie"
+require "rails/all"
 # require "rails/test_unit/railtie"
 
 Bundler.require(*Rails.groups)
@@ -9,17 +9,17 @@ require "liquid-rails"
 
 module Dummy
   class Application < Rails::Application
+    if Rails.version.start_with?('5.1')
+      # Initialize configuration defaults for originally generated Rails version.
+      config.load_defaults 5.1
+    elsif Rails.version.start_with?('5.2')
+      # Initialize configuration defaults for originally generated Rails version.
+      config.load_defaults 5.2
+      config.active_record.sqlite3.represent_boolean_as_integer = true
+    end
+
     # Settings in config/environments/* take precedence over those specified here.
     # Application configuration should go into files in config/initializers
     # -- all .rb files in that directory are automatically loaded.
-
-    # Set Time.zone default to the specified zone and make Active Record auto-convert to this zone.
-    # Run "rake -D time" for a list of tasks for finding time zone names. Default is UTC.
-    # config.time_zone = 'Central Time (US & Canada)'
-
-    # The default locale is :en and all translations from config/locales/*.rb,yml are auto loaded.
-    # config.i18n.load_path += Dir[Rails.root.join('my', 'locales', '*.{rb,yml}').to_s]
-    # config.i18n.default_locale = :de
   end
 end
-
