@@ -8,18 +8,18 @@ module Liquid
       context '#index' do
         it 'returns value at the specified index' do
           context['array'] = [1, 2, 3]
-          expect(::Liquid::Variable.new("array | index: 0").render(context)).to eq(1)
+          expect(::Liquid::Variable.new("array | index: 0", ParseContext.new).render(context)).to eq(1)
         end
 
         it 'returns nil when outside range' do
           context['array'] = [1, 2, 3]
-          expect(::Liquid::Variable.new("array | index: 5").render(context)).to eq(nil)
+          expect(::Liquid::Variable.new("array | index: 5", ParseContext.new).render(context)).to eq(nil)
         end
       end
 
       it '#jsonify' do
         context['listing'] = { name: 'Listing A' }
-        expect(::Liquid::Variable.new("listing | jsonify").render(context)).to eq(%|{"name":"Listing A"}|)
+        expect(::Liquid::Variable.new("listing | jsonify", ParseContext.new).render(context)).to eq(%|{"name":"Listing A"}|)
       end
     end
   end
