@@ -10,19 +10,19 @@ module Liquid
       end
 
       it 'translate with default locale' do
-        expect(::Liquid::Variable.new("'welcome' | translate").render(context)).to eq('Welcome everyone!')
+        expect(::Liquid::Variable.new("'welcome' | translate", ParseContext.new).render(context)).to eq('Welcome everyone!')
       end
 
       it 'translate with specified locale' do
-        expect(::Liquid::Variable.new("'welcome' | translate: locale: 'km'").render(context)).to eq('សូមស្វាគមន៍')
+        expect(::Liquid::Variable.new("'welcome' | translate: locale: 'km'", ParseContext.new).render(context)).to eq('សូមស្វាគមន៍')
       end
 
       it 'translate with scope' do
-        expect(::Liquid::Variable.new("'home' | translate: locale: 'km', scope: 'links'").render(context)).to eq('ទំព័រដើម')
+        expect(::Liquid::Variable.new("'home' | translate: locale: 'km', scope: 'links'", ParseContext.new).render(context)).to eq('ទំព័រដើម')
       end
 
       it 'translate with interpolation' do
-        expect(::Liquid::Variable.new("'welcome_name' | translate: locale: 'en', name: 'Jeremy'").render(context)).to eq('Welcome, Jeremy')
+        expect(::Liquid::Variable.new("'welcome_name' | translate: locale: 'en', name: 'Jeremy'", ParseContext.new).render(context)).to eq('Welcome, Jeremy')
       end
     end
   end
